Add files via upload
The program firstly parses the text file into words and store the words into a linked list of nodes, each node has a word and the count variable to store the frequencies. At the end, the program creates an array of array_nodes and use the copy() function to store the table into the array.